fixed translation issue and refactored folder structure
