+++ src/app/components/components.module.ts
@@ -2,6 +2,7 @@ import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
 import { FormsModule } from "@angular/forms";
 import { RouterModule } from "@angular/router";
+import { TranslateModule } from "@ngx-translate/core";
 import { ModalModule } from "ngx-bootstrap/modal";
 import { NgxSpinnerModule } from "ngx-spinner";
 import { PageNotFoundModule } from "./des-page-not-found/page-not-found.module";
@@ -9,7 +10,7 @@ import { COMPONENTS } from "./index";
 
 @NgModule({
 	declarations: [...COMPONENTS],
-	imports: [CommonModule, FormsModule, ModalModule.forRoot(), NgxSpinnerModule, PageNotFoundModule, RouterModule],
+	imports: [CommonModule, FormsModule, TranslateModule, ModalModule.forRoot(), NgxSpinnerModule, PageNotFoundModule, RouterModule],
 	exports: [...COMPONENTS]
 })
 export class ComponentsModule {}
