+++ src/app/services/auth/auth.service.ts
@@ -2,7 +2,7 @@ import { HttpClient } from "@angular/common/http";
 import { Injectable } from "@angular/core";
 import { Observable, delay } from "rxjs";
 import { environment, getEndpointUrl } from "src/environments/environment";
-import { API } from "../enums";
+import { API } from "../../enums";
 
 @Injectable({
 	providedIn: "root"
