+++ src/app/shared/shared.module.ts
@@ -1,12 +1,13 @@
 import { CommonModule } from "@angular/common";
-import { HttpClientModule } from "@angular/common/http";
+import { HttpClient, HttpClientModule } from "@angular/common/http";
 import { NgModule } from "@angular/core";
 import { FormsModule, ReactiveFormsModule } from "@angular/forms";
 import { MatDatepickerModule } from "@angular/material/datepicker";
 import { BrowserModule } from "@angular/platform-browser";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 import { RouterModule } from "@angular/router";
-import { TranslateModule } from "@ngx-translate/core";
+import { TranslateLoader, TranslateModule } from "@ngx-translate/core";
+import { TranslateHttpLoader } from "@ngx-translate/http-loader";
 import { ModalModule } from "ngx-bootstrap/modal";
 import { ProgressAnimationType, ToastrModule } from "ngx-toastr";
 import { ComponentsModule } from "src/app/components/components.module";
@@ -24,7 +25,13 @@ import { CONTAINERS } from "src/app/containers";
 		ReactiveFormsModule,
 		ModalModule.forRoot(),
 		ComponentsModule,
-		TranslateModule.forRoot(),
+		TranslateModule.forRoot({
+			loader: {
+				provide: TranslateLoader,
+				useFactory: (http: HttpClient) => new TranslateHttpLoader(http, "./assets/i18n/", ".json"),
+				deps: [HttpClient]
+			}
+		}),
 		ToastrModule.forRoot({
 			maxOpened: 1,
 			timeOut: 5000,
